+++ src/main/java/com/kl/grooveo/boundedContext/member/dto/JoinFormDTO.java
@@ -1,14 +1,14 @@
-package com.kl.grooveo.boundedContext.member.form;
+package com.kl.grooveo.boundedContext.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
-@AllArgsConstructor
-public class JoinForm {
+@Setter
+public class JoinFormDTO {
 
 	@NotBlank
 	@Size(min = 3, max = 30)
@@ -34,5 +34,4 @@ public class JoinForm {
 	@Email
 	@Size(min = 3, max = 30)
 	private String email;
-
 }
